Refactor support for a list of TAP queries
Rather than add a new layer of abstraction to allow two different
configurations of TAPQueryRunner, abstract the common functionality
of businesses that run TAP queries, rename the old TAPQueryRunner
to TAPQuerySetRunner, and add a new TAPQueryRunner that takes a
simple list of queries. This follows the existing pattern for keeping
each business simple and supporting a variety of tests via separate
business classes.
