+++ changelog.d/20231214_124422_rra_DM_42182.md
@@ -1,0 +1,3 @@
+### Backwards-incompatible changes
+
+- Rename the existing `TAPQueryRunner` business to `TAPQuerySetRunner` to more accurately capture what it does. Add a new `TAPQueryRunner` business that runs queries chosen randomly from a list. Based on work by @stvoutsin.
