+++ src/mobu/data/tapquerysetrunner/dp0.1/cone-photometry.sql
