+++ src/mobu/data/tapquerysetrunner/dp0.1/object-photometry.sql
