+++ src/mobu/data/tapquerysetrunner/dp0.1/scan-magnitude.sql
