+++ src/mobu/services/monkey.py
@@ -22,6 +22,7 @@ from ..models.business.empty import EmptyLoopConfig
 from ..models.business.jupyterpythonloop import JupyterPythonLoopConfig
 from ..models.business.notebookrunner import NotebookRunnerConfig
 from ..models.business.tapqueryrunner import TAPQueryRunnerConfig
+from ..models.business.tapquerysetrunner import TAPQuerySetRunnerConfig
 from ..models.monkey import MonkeyData, MonkeyState
 from ..models.user import AuthenticatedUser
 from .business.base import Business
@@ -29,6 +30,7 @@ from .business.empty import EmptyLoop
 from .business.jupyterpythonloop import JupyterPythonLoop
 from .business.notebookrunner import NotebookRunner
 from .business.tapqueryrunner import TAPQueryRunner
+from .business.tapquerysetrunner import TAPQuerySetRunner
 
 __all__ = ["Monkey"]
 
@@ -97,6 +99,10 @@ class Monkey:
             )
         elif isinstance(business_config, TAPQueryRunnerConfig):
             self.business = TAPQueryRunner(
+                business_config.options, user, self._http_client, self._logger
+            )
+        elif isinstance(business_config, TAPQuerySetRunnerConfig):
+            self.business = TAPQuerySetRunner(
                 business_config.options, user, self._http_client, self._logger
             )
         else:
